requirements.txt
# CORE DEPENDENCIES
PyMuPDF==1.22.5  # PDF text extraction
pdf2image==1.16.3  # PDF to image conversion
pytesseract==0.3.10  # OCR processing
Pillow==10.1.0  # Image handling
beautifulsoup4==4.12.0  # HTML parsing
requests==2.31.0  # HTTP client

# OPTIONAL (FOR EPUB SUPPORT)
ebooklib==0.18
PyPDF2==3.0.1

# DEVELOPMENT
python-dotenv==1.0.0  # Environment variables
tqdm==4.66.1  # Progress bars